Feature/tk 007 walked dogs

+++ src/main/java/com/petsuite/Services/repository/DogWalkerRepository.java
@@ -3,4 +3,4 @@ package com.petsuite.Services.repository;
 import com.petsuite.Services.model.DogWalker;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DogWalkerRepository extends JpaRepository<DogWalker, Integer>{}
+public interface DogWalkerRepository extends JpaRepository<DogWalker, Float>{}
